Made the `EmailSender` class use a separate username in addition to the sender email adderss.
