+++ api/sample.env
@@ -8,17 +8,10 @@ WZDX_API_KEY = <api-key>
 WZDX_ENDPOINT = <wzdx-endpoint>
 BSM_DB_NAME = <bsm visualization db name>
 
-# The email that the CV Manager will use to send emails
+# Contact Support Email Configuration
 EMAIL_TO_SEND_FROM=
-
-# The app password for the email that the CV Manager will use to send emails
+EMAIL_APP_USERNAME=
 EMAIL_APP_PASSWORD=
-
-# List of email addresses separated by commas that the CV Manager will send emails to
 EMAILS_TO_SEND_TO=
-
-# Target SMTP server for the CV Manager to send emails through
 TARGET_SMTP_SERVER_ADDRESS=
-
-# Target SMTP server port for the CV Manager to send emails through
 TARGET_SMTP_SERVER_PORT=
