+++ api/src/emailSender.py
@@ -8,12 +8,12 @@ class EmailSender():
         self.context = ssl._create_unverified_context()
         self.server = smtplib.SMTP(self.smtp_server, self.port)
     
-    def send(self, sender, recipient, subject, message, replyEmail, emailAppPassword):
+    def send(self, sender, recipient, subject, message, replyEmail, username, password):
         try:
             self.server.ehlo() # say hello to server
             self.server.starttls(context=self.context) # start TLS encryption
             self.server.ehlo() # say hello again
-            self.server.login(sender, emailAppPassword)
+            self.server.login(username, password)
 
             # prepare email
             toSend = self.prepareEmailToSend(sender, recipient, subject, message, replyEmail)
