+++ api/tests/data/contact_support_data.py
@@ -1,5 +1,6 @@
 EMAIL_TO_SEND_FROM = "test@test.com"
 EMAILS_TO_SEND_TO = "bob@test.com"
+EMAIL_APP_USERNAME = "testusername"
 EMAIL_APP_PASSWORD = "testpassword"
 
 EMAIL_REPLY_EMAIL = "reply-to@test.com"
