+++ api/tests/src/test_emailSender.py
@@ -1,5 +1,5 @@
 from unittest.mock import MagicMock
-from api.src.helpers.emailSender import EmailSender
+from api.src.emailSender import EmailSender
 
 
 EMAIL_TO_SEND_FROM = "test@test.test"
@@ -7,6 +7,7 @@ EMAILS_TO_SEND_TO = ["test@test.test, test2@test.test"]
 EMAIL_SUBJECT = "Test Subject"
 EMAIL_MESSAGE = "Test Message"
 EMAIL_REPLY_EMAIL = "test@test.test"
+EMAIL_APP_USERNAME = "test"
 EMAIL_APP_PASSWORD = "test"
 DEFAULT_TARGET_SMTP_SERVER_ADDRESS = "smtp.gmail.com"
 DEFAULT_TARGET_SMTP_SERVER_PORT = 587
@@ -22,7 +23,7 @@ def test_send():
     emailSender.server.quit = MagicMock()
 
     # execute
-    emailSender.send(EMAIL_TO_SEND_FROM, EMAILS_TO_SEND_TO, EMAIL_SUBJECT, EMAIL_MESSAGE, EMAIL_REPLY_EMAIL, EMAIL_APP_PASSWORD)
+    emailSender.send(EMAIL_TO_SEND_FROM, EMAILS_TO_SEND_TO, EMAIL_SUBJECT, EMAIL_MESSAGE, EMAIL_REPLY_EMAIL, EMAIL_APP_USERNAME, EMAIL_APP_PASSWORD)
 
     # assert
     emailSender.server.starttls.assert_called_once()
